nnUNet-msgpu1.10/requirements.txt
tqdm
dicom2nifti
scikit-image>=0.14
medpy
scipy
batchgenerators>=0.23
numpy
scikit-learn
SimpleITK
pandas
requests
nibabel
tifffile
pillow==9.5.0
matplotlib
mindspore-gpu==1.10.0
